feat(ui): Introducing reselect selectors (AEROGEAR-8922)
