+++ ui/package.json
@@ -19,7 +19,8 @@
     "react-scripts": "2.1.3",
     "redux": "4.0.1",
     "redux-logger": "3.0.6",
-    "redux-thunk": "2.3.0"
+    "redux-thunk": "2.3.0",
+    "reselect": "^4.0.0"
   },
   "scripts": {
     "start": "react-scripts start",
